requirements.txt
aws-cdk.aws-bedrock-agentcore-alpha
aws-cdk-lib>=2.200.0
constructs>=10.0.0,<11.0.0
boto3
botocore
requests
mcp